Separate grid implementations into separate files
